Back to pre-build packages, update to v. 2.15.1
Listing all dependencies, trying to reproduce on node/npm was giving a
lot of issues. Since it was picking electron 4.0.8, there was a lot of
issues with symbols of compiled libraries as well, which failed
completely to load even on same-version Node. The build size is smaller
on the released package as well.