Update to v2.15.5
Unfortunately, upstream removed i386 builds, and the overhead to provide
built-from-source packages is just too high now (with the whole
dependency pinning thing), so from now on, only x86_64 builds.
If anyone finds a simpler way to get an Electron package with tons of
deps easily built and updateable in flatpak, I'll be glad to review PRs.